+++ tests/test_terachem.py
@@ -1,25 +1,17 @@
-from pathlib import Path
-
 import pytest
 from qcio import CalcType
 
 from qcparse.exceptions import MatchNotFoundError
 from qcparse.parsers.terachem import (
     calculation_succeeded,
-    parse_basis,
     parse_calctype,
     parse_energy,
     parse_git_commit,
     parse_gradient,
     parse_hessian,
-    parse_method,
-    parse_molecule_charge,
-    parse_molecule_spin_multiplicity,
     parse_natoms,
     parse_nmo,
     parse_version,
-    parse_working_directory,
-    parse_xyz_filepath,
 )
 
 from .data import gradients, hessians
@@ -42,7 +34,7 @@ def test_parse_energy(test_data_dir, data_collector, filename, energy):
 def test_parse_energy_positive(data_collector):
     energy = 76.3854579982
     parse_energy(f"FINAL ENERGY: {energy} a.u", data_collector)
-    assert data_collector.computed.energy == energy
+    assert data_collector.energy == energy
 
 
 @pytest.mark.parametrize(
@@ -51,7 +43,7 @@ def test_parse_energy_positive(data_collector):
 )
 def test_parse_energy_integer(data_collector, energy):
     parse_energy(f"FINAL ENERGY: {energy} a.u", data_collector)
-    assert data_collector.computed.energy == energy
+    assert data_collector.energy == energy
 
 
 def test_parse_energy_raises_exception(data_collector):
@@ -78,50 +70,12 @@ def test_parse_calctype_raises_exception():
         parse_calctype("No driver here")
 
 
-@pytest.mark.parametrize(
-    "string,path",
-    (
-        ("XYZ coordinates water.xyz", Path("water.xyz")),
-        ("XYZ coordinates water", Path("water")),
-        ("XYZ coordinates /scratch/water.xyz", Path("/scratch/water.xyz")),
-        ("XYZ coordinates ../water.xyz", Path("../water.xyz")),
-    ),
-)
-def test_parse_xyz_filepath(string, path):
-    assert parse_xyz_filepath(string) == path
-
-
-@pytest.mark.parametrize(
-    "filename,method",
-    (
-        ("water.energy.out", "B3LYP"),
-        ("water.gradient.out", "B3LYP"),
-        ("water.frequencies.out", "B3LYP"),
-    ),
-)
-def test_parse_method(test_data_dir, filename, method, data_collector):
-    with open(test_data_dir / filename) as f:
-        tcout = f.read()
-    parse_method(tcout, data_collector)
-    assert data_collector.input_data.model.method == method
-
-
-def test_parse_basis(terachem_energy_stdout, data_collector):
-    parse_basis(terachem_energy_stdout, data_collector)
-    assert data_collector.input_data.model.basis == "6-31g"
-
-
 def test_parse_version(terachem_energy_stdout, data_collector):
     parse_version(terachem_energy_stdout, data_collector)
     assert (
-        data_collector.provenance.program_version
+        data_collector.extras.program_version
         == "v1.9-2022.03-dev [4daa16dd21e78d64be5415f7663c3d7c2785203c]"
     )
-
-
-def test_parse_working_dir(terachem_energy_stdout, data_collector):
-    parse_working_directory(terachem_energy_stdout, data_collector)
-    assert data_collector.provenance.working_dir == "./scr.water"
 
 
 def test_calculation_succeeded(terachem_energy_stdout):
@@ -173,7 +127,7 @@ def test_parse_gradient(test_data_dir, filename, gradient, data_collector):
         tcout = f.read()
 
     parse_gradient(tcout, data_collector)
-    assert data_collector.computed.gradient == gradient
+    assert data_collector.gradient == gradient
 
 
 @pytest.mark.parametrize(
@@ -194,7 +148,7 @@ def test_parse_hessian(test_data_dir, filename, hessian, data_collector):
         tcout = f.read()
 
     parse_hessian(tcout, data_collector)
-    assert data_collector.computed.hessian == hessian
+    assert data_collector.hessian == hessian
 
 
 @pytest.mark.parametrize(
@@ -206,7 +160,7 @@ def test_parse_natoms(test_data_dir, filename, n_atoms, data_collector):
         tcout = f.read()
 
     parse_natoms(tcout, data_collector)
-    assert data_collector.computed.calcinfo_natoms == n_atoms
+    assert data_collector.calcinfo_natoms == n_atoms
 
 
 @pytest.mark.parametrize(
@@ -218,29 +172,7 @@ def test_parse_nmo(test_data_dir, filename, nmo, data_collector):
         tcout = f.read()
 
     parse_nmo(tcout, data_collector)
-    assert data_collector.computed.calcinfo_nmo == nmo
-
-
-@pytest.mark.parametrize(
-    "filename,charge",
-    (("water.energy.out", 0), ("caffeine.gradient.out", 0)),
-)
-def test_parse_molecule_charge(test_data_dir, filename, charge):
-    with open(test_data_dir / filename) as f:
-        tcout = f.read()
-    n = parse_molecule_charge(tcout)
-    assert n == charge
-
-
-@pytest.mark.parametrize(
-    "filename,multiplicity",
-    (("water.energy.out", 1), ("caffeine.gradient.out", 1)),
-)
-def test_parse_molecule_spin_multiplicity(test_data_dir, filename, multiplicity):
-    with open(test_data_dir / filename) as f:
-        tcout = f.read()
-    n = parse_molecule_spin_multiplicity(tcout)
-    assert n == multiplicity
+    assert data_collector.calcinfo_nmo == nmo
 
 
 def test_parse_git_commit(terachem_energy_stdout):